pyproject.toml
[tool.poetry]
name = "merkly"
version = "0.3.0"
readme = "README.md"
description = "🌳 The simple and easy implementation of Merkle Tree"
authors = ["Lucas Oliveira <olivmath@protonmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
pysha3 = "^1.0.2"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
